chore: renamed modules to canonical source (#22)
* chore: renamed modules to canonical source

Module path was devrel/cloud/devrel-github-services is now
github.com/GoogleCloudPlatform/devrel-services

* fix: updated tests for drghs-worker
