+++ drghs-worker/maintnerd/internal/apiroutes/v1_api.go
@@ -15,9 +15,9 @@
 package apiroutes
 
 import (
-	"devrel/cloud/devrel-github-service/drghs-worker/pkg/googlers"
-	"devrel/cloud/devrel-github-service/drghs-worker/pkg/status"
-	"devrel/cloud/devrel-github-service/drghs-worker/pkg/utils"
+	"github.com/GoogleCloudPlatform/devrel-services/drghs-worker/pkg/googlers"
+	"github.com/GoogleCloudPlatform/devrel-services/drghs-worker/pkg/status"
+	"github.com/GoogleCloudPlatform/devrel-services/drghs-worker/pkg/utils"
 	"fmt"
 	"net/http"
 	"strconv"
