+++ drghs/go.mod
@@ -15,7 +15,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // source: issue_service.proto
 
-module devrel/cloud/devrel-github-service/drghs
+module github.com/GoogleCloudPlatform/devrel-services/drghs
 
 require (
 	github.com/golang/protobuf v1.2.0
