+++ samplr/git-go/go.mod
@@ -1,4 +1,4 @@
-module devrel/cloud/devrel-github-service/git-go
+module github.com/GoogleCloudPlatform/devrel-services/git-go
 
 require github.com/sirupsen/logrus v1.3.0
