+++ samplr/samplrctl/snippets/snippets.go
@@ -18,8 +18,8 @@ import (
 	"io"
 	"strconv"
 
-	"devrel/cloud/devrel-github-service/samplr"
-	"devrel/cloud/devrel-github-service/samplr/samplrctl/output"
+	"github.com/GoogleCloudPlatform/devrel-services/samplr"
+	"github.com/GoogleCloudPlatform/devrel-services/samplr/samplrctl/output"
 )
 
 const (
