+++ samplr/samplrd/samplrapi/sample_service.go
@@ -22,9 +22,9 @@ import (
 	"regexp"
 	"time"
 
-	drghs_v1 "devrel/cloud/devrel-github-service/drghs/v1"
-	"devrel/cloud/devrel-github-service/samplr"
-	"devrel/cloud/devrel-github-service/samplr/samplrd/filter"
+	drghs_v1 "github.com/GoogleCloudPlatform/devrel-services/drghs/v1"
+	"github.com/GoogleCloudPlatform/devrel-services/samplr"
+	"github.com/GoogleCloudPlatform/devrel-services/samplr/samplrd/filter"
 
 	"github.com/golang/protobuf/ptypes"
 	"github.com/sirupsen/logrus"
