+++ samplr/samplrd/samplrapi/snippet_paginator_test.go
@@ -15,7 +15,7 @@
 package samplrapi
 
 import (
-	"devrel/cloud/devrel-github-service/samplr"
+	"github.com/GoogleCloudPlatform/devrel-services/samplr"
 	"reflect"
 	"testing"
 	"time"
